FIX: Regression in path for uploads on s3.

+++ app/jobs/regular/backup_upload_to_s3.rb
@@ -3,28 +3,30 @@ module Jobs
     def execute(args)
       upload = Upload.find(args[:upload_id])
 
-      if path = "#{Discourse.store.path_for(upload)}.gpg"
+      if local_path = Discourse.store.path_for(upload)
         s3_gnupg_public_key = GlobalSetting.backup_uploads_to_s3_gnupg_public_key
-        tmp_path = "/tmp/#{File.basename(path)}"
+        encrypted_filename = "#{File.basename(local_path)}.gpg"
+        tmp_path = "/tmp/#{encrypted_filename}"
 
         `echo '#{s3_gnupg_public_key}' | gpg --import`
-        gnupg_user_id = `echo '#{s3_gnupg_public_key}' | gpg --list-keys --with-colons | awk -F: '/^pub:/ { print $5 }'`.chomp
+        gnupg_user_id = `gpg --list-keys --with-colons | awk -F: '/^pub:/ { print $5 }'`.chomp
 
         begin
-          `gpg --encrypt --output #{tmp_path} --batch --yes --verbose --recipient #{gnupg_user_id} #{@path}`
-          File.open(tmp_path) { |file| backup_upload(file, path, upload.id) }
+          `gpg --encrypt --output #{tmp_path} --batch --yes --verbose --recipient #{gnupg_user_id} #{local_path}`
+          File.open(tmp_path) { |file| backup_upload(file, upload) }
         ensure
-          File.delete(tmp_path)
+          File.delete(tmp_path) if File.exists?(tmp_path)
         end
       end
     end
 
-    def backup_upload(file, path, upload_id)
+    def backup_upload(file, upload)
+      path = "#{DiscourseBackupUploadsToS3::Utils.s3_store.get_path_for_upload(upload)}.gpg"
       path = DiscourseBackupUploadsToS3::Utils.s3_helper.upload(file, path)
 
       PluginStore.set(
         DiscourseBackupUploadsToS3::PLUGIN_NAME,
-        DiscourseBackupUploadsToS3::Utils.plugin_store_key(upload_id),
+        DiscourseBackupUploadsToS3::Utils.plugin_store_key(upload.id),
         "#{DiscourseBackupUploadsToS3::Utils.s3_store.absolute_base_url}/#{path}"
       )
     end
